fix: improve behavior of AutoLogin._check_network
The captive portal of China Telecom APs does not function as expected
sometimes with the presence of a transparent proxy. When that happens,
all the http requests are responded with a HTTP 502, including the ones
supposed to be redirected to the login page. Besides, determining
internet connectivity like this is still slow and unreliable.

It is recommended [1] to use an IP (to avoid DNS resolution) of a
service with high SLA (like 1.1.1.1).

---
[1] https://stackoverflow.com/a/33117579