migrate dockerfile to official python image
changes:
- use multistage build for seperating slim / full build
- use python:3.11-alpine as base image

Ubuntu is too 'heavy' for projects like this, and alpine is enough. This
change will bring the size of the image down to dozens of MB (for slim
build) or hundreds of MB (for full build), and allow it to be run on
devices with poorer performance. Note that although the build can pass
and run, test in production environment is still needed.